extractors/okru.py
import json
import logging
import random
from aiohttp import ClientSession, ClientTimeout, TCPConnector
from aiohttp_socks import ProxyConnector
from bs4 import BeautifulSoup, SoupStrainer

logger = logging.getLogger(__name__)

class ExtractorError(Exception):
    pass

class OkruExtractor:
    """Okru (ok.ru) URL extractor."""

    def __init__(self, request_headers: dict, proxies: list = None):
        self.request_headers = request_headers
        self.base_headers = {
            "user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
        }
        self.session = None
        self.mediaflow_endpoint = "hls_proxy"
        self.proxies = proxies or []

    def _get_random_proxy(self):
        return random.choice(self.proxies) if self.proxies else None

    async def _get_session(self):
        if self.session is None or self.session.closed:
            timeout = ClientTimeout(total=60, connect=30, sock_read=30)
            proxy = self._get_random_proxy()
            if proxy:
                connector = ProxyConnector.from_url(proxy)
            else:
                connector = TCPConnector(limit=0, limit_per_host=0, keepalive_timeout=60, enable_cleanup_closed=True, force_close=False, use_dns_cache=True)
            self.session = ClientSession(timeout=timeout, connector=connector, headers={'User-Agent': self.base_headers["user-agent"]})
        return self.session

    async def extract(self, url: str, **kwargs) -> dict:
        """Extract Okru URL."""
        session = await self._get_session()
        
        async with session.get(url) as response:
            text = await response.text()

        soup = BeautifulSoup(text, "lxml", parse_only=SoupStrainer("div"))
        if soup:
            div = soup.find("div", {"data-module": "OKVideo"})
            if not div:
                raise ExtractorError("Failed to find video element")
            
            data_options = div.get("data-options")
            data = json.loads(data_options)
            metadata = json.loads(data["flashvars"]["metadata"])
            final_url = (
                metadata.get("hlsMasterPlaylistUrl") or metadata.get("hlsManifestUrl") or metadata.get("ondemandHls")
            )
            
            if not final_url:
                raise ExtractorError("Failed to extract stream URL from metadata")
            
            self.base_headers["referer"] = url
            return {
                "destination_url": final_url,
                "request_headers": self.base_headers,
                "mediaflow_endpoint": self.mediaflow_endpoint,
            }
        
        raise ExtractorError("Failed to parse OK.ru page")

    async def close(self):
        if self.session and not self.session.closed:
            await self.session.close()
