requirements.txt
aiohttp
aiohttp-socks
gunicorn
python-dotenv
zstandard
beautifulsoup4
lxml
python-socks
pycryptodome
pydash2hls
